// src/app/public/models/ROM/entel/validacion-bundles/validacionbundle.ts
export class ValidacionBundle{
    idventas?: number;
    idventasdetalle?: number;
    fechaoperacion?: string;
    doccliente?: string;
    numcelularcontrato?: string;
    docpromotorasesor?: string;
    nombrepromotor?: string;
    // dnipromotor?: string;
    idsubproducto?: number;
    nombresubproducto?: string;
    idplan?: number;
    nombreplan?: string;
    idmodelo?: number;
    nombremodelo?: string;
    idbundle?: number;
    codigobundle?: string;
    nombrebundle?: string;
    nombreobsequio?: string;
    idemppaisnegcue?: number;
    nombrepdffirma?: string;
    codigoauthbundle?: string;
    fechacreacion?: string;
    numeroorden?: string;
    idpdv?: number;
    nombrepdv?: string;
    flagcodigoauthbundle?: number;
    usuariocreacion?: string;
}

export class ValidacionBundleROMWEB{
    intidventasprincipal?: number;
    intventasromid?: number;
    strdtevestasromfeope?: string;
    strdnicliente?: string;
    strcelularcontrato?: string;
    strventasromusucr?: string;
    nombrepromotor?: string;
    dnipromotor?: string;
    intproductoid?: number;
    strproductodesc?: string;
    intplanid?: number;
    strplandesc?: string;
    intmodeloequipoid?: number;
    strmodeloequipodesc?: string;
    intbundleid?: number;
    codigo?: string;
    descripcion?: string;
    nombreobsequio?: string;
    nombrepdffirma?: string;
    strcodigoauthbundle?: string;
    dteventasromfecre?: string;
    strnumorden?: string;
    idpuntoventa?: number;
    puntoventa?: string;
    flagcodigoauthbundle?: number;
    usuario_creacion?: string;
}